Added DB file. Updated README. Changed Create React App (CRA) to Vite.js because of dependency errors.
